Updated spam command
Only people with admin permissions in a server can use spam command. Changed it because I got some issues with random people using spam command to spam unlimited messages to a server hence causing a havoc. That's why the bot was offline these days.  Also need to create a command that can stop the spam command in future.